add delete grace feature

+++ cmd/main.go
@@ -4,13 +4,15 @@ import (
 	"fmt"
 	"github.com/cloudflare/cloudflare-go"
 	"os"
+	"strconv"
+	"strings"
 
 	internal "github.com/containeroo/syncflaer/internal"
 
 	log "github.com/sirupsen/logrus"
 )
 
-const version string = "1.0.6"
+const version string = "1.1.0"
 
 func main() {
 	log.SetOutput(os.Stdout)
@@ -33,14 +35,14 @@ func main() {
 
 	log.Debugf("SyncFlaer %s", version)
 
-	internal.GetConfig(configFilePath)
+	config := internal.GetConfig(configFilePath)
 
 	internal.SetupCloudflareClient()
 	internal.GetCloudflareZoneID()
 
 	internal.GetCurrentIP()
 
-	cloudflareDNSRecords := internal.GetCloudflareDNSRecords()
+	cloudflareDNSRecords, deleteGraceRecords := internal.GetCloudflareDNSRecords()
 
 	var userRecords []cloudflare.DNSRecord
 	userRecords = internal.GetTraefikRules(userRecords)
@@ -58,15 +60,50 @@ func main() {
 	orphanedRecords := internal.GetOrphanedDNSRecords(cloudflareDNSRecords, userRecords)
 	if orphanedRecords != nil {
 		for _, orphanedRecord := range orphanedRecords {
-			internal.DeleteCloudflareDNSRecord(orphanedRecord)
+			if config.Cloudflare.DeleteGrace == 0 {
+				internal.DeleteCloudflareDNSRecord(orphanedRecord)
+				for _, deleteGraceRecord := range deleteGraceRecords {
+					log.Infof("Cleaning up delete grace DNS record for %s", strings.TrimPrefix(deleteGraceRecord.Name, "_syncflaer._deletegrace."))
+					internal.DeleteCloudflareDNSRecord(deleteGraceRecord)
+				}
+				continue
+			}
+			deleteGraceRecordFound := false
+			for _, deleteGraceRecord := range deleteGraceRecords {
+				if !strings.Contains(deleteGraceRecord.Name, orphanedRecord.Name) {
+					continue
+				}
+				deleteGraceRecordFound = true
+				deleteGrace, _ := strconv.Atoi(deleteGraceRecord.Content)
+				deleteGrace -= 1
+				if deleteGrace > 0 {
+					deleteGraceRecord.Content = strconv.Itoa(deleteGrace)
+					log.Infof("Waiting %d more runs until deleting DNS record %s", deleteGrace-1, orphanedRecord.Name)
+					internal.UpdateCloudflareDNSRecord(deleteGraceRecord)
+					continue
+				}
+				internal.DeleteCloudflareDNSRecord(orphanedRecord)
+				internal.DeleteCloudflareDNSRecord(deleteGraceRecord)
+			}
+			if !deleteGraceRecordFound {
+				deleteGraceRecord := cloudflare.DNSRecord{
+					Type:    "TXT",
+					Name:    fmt.Sprintf("_syncflaer._deletegrace.%s", orphanedRecord.Name),
+					Content: strconv.Itoa(config.Cloudflare.DeleteGrace),
+				}
+				log.Infof("Waiting %d more runs until deleting DNS record %s", config.Cloudflare.DeleteGrace-1, orphanedRecord.Name)
+				internal.CreateCloudflareDNSRecord(deleteGraceRecord)
+			}
 		}
 	} else {
+		for _, deleteGraceRecord := range deleteGraceRecords {
+			log.Infof("DNS record %s is not orphaned anymore", strings.TrimPrefix(deleteGraceRecord.Name, "_syncflaer._deletegrace."))
+			internal.DeleteCloudflareDNSRecord(deleteGraceRecord)
+		}
 		log.Debug("No orphaned DNS records")
 	}
 
-	internal.UpdateCloudflareDNSRecords(cloudflareDNSRecords, userRecords)
+	internal.UpdateOutdatedDNSRecords(cloudflareDNSRecords, userRecords)
 
 	internal.SendSlackMessage()
-
-	log.Debug("Done")
 }
